pyqt5/requirements.txt
PyQt5
pyzmq
txzmq
